Only use expectedDsoParty in tests when it is available
Fixes https://github.com/DACH-NY/cn-test-failures/issues/4847

[ci]
